Tighten inference module names
